move other devices in component
